+++ src/operand.h
@@ -58,7 +58,7 @@ class Operand : public Token {
         }
 
         static bool isOperand(const string & token) {
-            for (int i = 0;i < token.length();i++) {
+            for (size_t i = 0;i < token.length();i++) {
                 char ch = token[i];
 
                 if (!isOperandChar(ch)) {
